Improving bfp in alu

+++ simulator/func_sim/alu.h
@@ -322,7 +322,7 @@ struct ALU
         size_t XLEN = bitwidth<XLENType>;
         size_t len = ( narrow_cast<size_t>( instr->v_src2) >> 24) & 15;
         len = len ? len : 16;
-        size_t off = ( instr->v_src2 >> 16) & ( XLEN-1);
+        size_t off = ( narrow_cast<size_t>( instr->v_src2) >> 16) & ( XLEN-1);
         auto mask = circ_ls( bitmask<XLENType>( len), off);
         auto data = circ_ls( instr->v_src2, off);
         instr->v_dst = ( data & mask) | ( instr->v_src1 & ~mask);
